Add extra optional argument for functions.
This change will add the ability to change the database document when calling a function.
For example if fetch_energy_data_by_date is called with the extra parameter db then the default environment variable "db" in the clean.cfg will be ignored.
if fetch_energy_data_by_date is not called with the extra parameter "db" then the default document name will be used from clean.cfg